chore: address issue #158
