+++ packages/seisai-engine/src/seisai_engine/pipelines/fbpick/physics/__init__.py
@@ -1,9 +1,11 @@
+"""Public exports for fbpick physics helpers."""
+
 from .config import (
     DEFAULT_PHYSICS_LITE_CONFIG,
     NeighborContextCfg,
+    PhysicalAdaptiveRefitCfg,
     PhysicalAnchorReuseCfg,
     PhysicalAnchorSelectionCfg,
-    PhysicalAdaptiveRefitCfg,
     PhysicalObservationSamplingCfg,
     PhysicalPrefilterCfg,
     PhysicalProgressCfg,
@@ -56,16 +58,18 @@ from .physical_center import (
     PHYSICAL_OFFSET_SOURCE_HEADER,
     PHYSICAL_OFFSET_SOURCE_LABELS,
     PHYSICAL_OFFSET_SOURCE_NONE,
-    PHYSICAL_RUNTIME_FIT_SOURCE_ANCHOR_FIT,
     PHYSICAL_RUNTIME_FIT_SOURCE_ADAPTIVE_REFIT,
+    PHYSICAL_RUNTIME_FIT_SOURCE_ANCHOR_FIT,
     PHYSICAL_RUNTIME_FIT_SOURCE_FALLBACK_EXISTING_TREND,
     PHYSICAL_RUNTIME_FIT_SOURCE_FALLBACK_FULL_FIT_NO_COMPATIBLE_ANCHOR,
     PHYSICAL_RUNTIME_FIT_SOURCE_FALLBACK_ROBUST,
     PHYSICAL_RUNTIME_FIT_SOURCE_FULL_FIT,
     PHYSICAL_RUNTIME_FIT_SOURCE_LABELS,
     PHYSICAL_RUNTIME_FIT_SOURCE_NEAREST_ANCHOR_REUSE,
+    PhysicalCenterFallbackPreflight,
     PhysicalCenterResult,
     build_geometry_two_piece_physical_center,
+    preflight_geometry_two_piece_fallback,
 )
 from .pick_table import CoarsePickTable, normalize_coarse_pick_table
 from .run import (
@@ -104,8 +108,8 @@ __all__ = [
     'PHYSICAL_OFFSET_SOURCE_HEADER',
     'PHYSICAL_OFFSET_SOURCE_LABELS',
     'PHYSICAL_OFFSET_SOURCE_NONE',
-    'PHYSICAL_RUNTIME_FIT_SOURCE_ANCHOR_FIT',
     'PHYSICAL_RUNTIME_FIT_SOURCE_ADAPTIVE_REFIT',
+    'PHYSICAL_RUNTIME_FIT_SOURCE_ANCHOR_FIT',
     'PHYSICAL_RUNTIME_FIT_SOURCE_FALLBACK_EXISTING_TREND',
     'PHYSICAL_RUNTIME_FIT_SOURCE_FALLBACK_FULL_FIT_NO_COMPATIBLE_ANCHOR',
     'PHYSICAL_RUNTIME_FIT_SOURCE_FALLBACK_ROBUST',
@@ -115,9 +119,10 @@ __all__ = [
     'CoarseGeometry',
     'CoarsePickTable',
     'NeighborContextCfg',
+    'PhysicalAdaptiveRefitCfg',
     'PhysicalAnchorReuseCfg',
     'PhysicalAnchorSelectionCfg',
-    'PhysicalAdaptiveRefitCfg',
+    'PhysicalCenterFallbackPreflight',
     'PhysicalCenterResult',
     'PhysicalObservationSamplingCfg',
     'PhysicalPrefilterCfg',
@@ -150,6 +155,7 @@ __all__ = [
     'normalize_coarse_pick_table',
     'order_source_groups_source_xy',
     'physics_lite_config_to_dict',
+    'preflight_geometry_two_piece_fallback',
     'run_physics_lite',
     'select_nearest_source_groups',
     'select_source_xy_stride_anchors',
